+++ mods/BrewLAN/documentation/Wiki Data.lua
@@ -1,4 +1,10 @@
 UnitData = {
+    SAA0105 = {
+        Trivia = {
+            'The "Respirer" was the second unit ever added to *BrewLAN*; added in August 2009.',
+            [["Respirer" had it's model replaced in July 2018, nearly 9 years after its creation.]],
+        },
+    },
     SAA0313 = {
         Videos = {
             {YouTube = 'CAVRJWb_acI', 'Time-lapse'},
@@ -24,6 +30,9 @@ UnitData = {
             {YouTube = 'BxmnEIWuX1s', 'Time-lapse'},
         },
     },
+    SAB0401 = {
+        Trivia = 'The design of the "Independance Engine" was inspired by a gas holder near the childhood home of the creator. The name is a reference to the 1996 film *Independence Day*.',
+    },
     SAB1102 = {
         Videos = {
             {YouTube = 'GzGiDBaemjI', 'Time-lapse'},
@@ -39,6 +48,9 @@ UnitData = {
             {YouTube = 'AHBsSCGcj98', 'Time-lapse'},
         },
     },
+    SAB1311 = {
+        Trivia = 'Julius von Mayer, for whom this unit was named, was a German scientist and one of the founders of thermodynamics.',
+    },
     SAB2221 = {
         Videos = {
             {YouTube = 'Wb6zP5j91y0', 'Time-lapse'},
@@ -49,10 +61,29 @@ UnitData = {
             {YouTube = 'Wb6zP5j91y0', 'Time-lapse'},
         },
     },
+    SAB2306 = {
+        Trivia = {
+            [[The name "Orbos" is based on a misremembering of the *Zelda* enemy 'Beamos', which has some cosmetic similarities.]],
+            [[This would have been called "Oculus", however that name was already taken by another tech 3 Aeon structure.]],
+        },
+    },
     SAB3301 = {
+        Trivia = '"Heimdallr", in Norse mythology, is a god known for their keep senses, especially eyesight and hearing.',
         Videos = {
             {YouTube = 'Dknhw6LYyDs', 'Time-lapse'},
         },
+    },
+    SAL0209 = {
+        Trivia = '"Butler" is named after two of the first testers of the mod.',
+    },
+    SAL0311 = {
+        Trivia = '"Moldavite" is named after a glassy green rock.',
+    },
+    SAL0319 = {
+        Trivia = '"Bilmon" is named after an early contributor and tester of the mod.',
+    },
+    SAL0321 = {
+        Trivia = '"Scarab" was the first walking unit added to *BrewLAN*; added in March 2014. It is loosely based on the *Total Annihilation* unit of the same name.',
     },
     SAL0323 = {
         Videos = {
@@ -64,44 +95,81 @@ UnitData = {
             {YouTube = 'OIYD1qzJU0U', 'Time-lapse'},
         },
     },
+    SAL0401 = {
+        Trivia = 'The "Absolution" was designed after having seen a Fire Prism one time.',
+    },
     SAS0401 = {
         Videos = {
             {YouTube = 'hOk93CkyZ2w', 'Time-lapse'},
+        },
+        Trivia = {
+            '"Deluge Class" was a commissioned unit.',
+            'The weapon layout of the Deluge Class is asymmetrical, the orientation of said weapons alternating with each one built.',
+            "The first two built in any given name will have the unique names 'Scylla' and 'Charybdis'.",
+            "If 69 Deluge Class ships are built in a given game, the 69th will be naked and have the unique name 'Indulge Class'.",
         },
     },
 
 
     SEA0105 = {
+        Trivia = {
+            'This was the first unit added to *BrewLAN*, added in August 2009.',
+            'In July 2019, almost 10 years after first creation, it was given a new model.',
+            'The original name for this unit was "Pidgeon", a subtle Pokémon reference. It was changed because people kept saying it was a typo.',
+        },
         Videos = {
             {YouTube = 'p8abzlcyFdg', 'Time-lapse'},
         },
     },
     SEA0201 = {
+        Trivia = 'The name and design of this are a UEF-themed interpretation of the Lockheed U-2 Dragon Lady.',
         Videos = {
             {YouTube = 'aDtB6VqyzrU', 'Time-lapse'},
         },
     },
-    SEA0212 = {
-        Videos = {
-            {YouTube = 'sFdvsuhXfaM', 'Time-lapse'},
-        },
+    SEA0211 = {
+        Trivia = '"Tornado" is based on the early jet-powered bomber the B-45 Tornado.',
     },
     SEA0212 = {
         Videos = {
             {YouTube = 'sFdvsuhXfaM', 'Time-lapse'},
         },
+        Trivia = '"Raptor" is based on the Lockheed Martin F-22 Raptor.',
     },
     SEA0307 = {
         Videos = {
             {YouTube = 'K_fPIx_Lt-M', 'Time-lapse'},
         },
     },
+    SEA0313 = {
+        Trivia = '"Tomcat" is based on the Grumman F-14 Tomcat, which appeared in the 1986 film *Top Gun*.',
+    },
     SEA0314 = {
         Videos = {
             {YouTube = 'fPAXxbB7nTg', 'Time-lapse'},
         },
+        Trivia = '"Lancer" is based on the Rockwell B-1B Lancer.',
+    },
+    SEA3301 = {
+        Trivia = [[The name "Burning Star" is a reference to the Coheed and Cambria album *Good Apollo, I'm Burning Star IV, Volume One: From Fear Through the Eyes of Madness*.]],
+    },
+    SEA4301 = {
+        Trivia = [[The name "Vaxis" is a reference to the Coheed and Cambria album *Vaxis – Act I: The Unheavenly Creatures*.]],
+    },
+    SEB1311 = {
+        Trivia = 'Sir Henry Bessemer, for whom this unit was named, invented a mass production process for steel. The basic concept of the unit came from a UEF upgrade on *Supreme Commander 2*.',
+    },
+    SEB2103 = {
+        Trivia = [["Poker" was the third unit added to *BrewLAN*; added in September 2009. It had its model replaced in June 2019, nearly 10 years later.]],
+    },
+    SEB2311 = {
+        Trivia = {
+            '"Guardian" is based on the *Total Annihilation* unit of the same name.',
+            'The name "Guardian" is not unique; it is shared by the Cybran strategic missile defense.',
+        },
     },
     SEB2401 = {
+        Trivia = '"Excalibur" was named by community poll. Runner up names were Maelstrom Cannon, Hyperion, Annihilator, Penetrator, and Daddy. Maelstrom was later used for the [Experimental Anti-Air Defense](SEB2402).',
         Videos = {
             {YouTube = '_I_D6KGNB8g', 'Time-lapse part 1'},
             {YouTube = 'rxOAF6QvWTc', 'Time-lapse part 2'},
@@ -112,12 +180,30 @@ UnitData = {
             {YouTube = 'ujHUtISdARI', 'Time-lapse'},
         },
     },
+    SEB2404 = {
+        Trivia = [["Ivan" is named after a friend of the creator who went by the name 'Noah' online; making it an indirect reference to the similar *Supreme Commander 2* unit.]],
+    },
+    SEB3301 = {
+        Trivia = '"Welchman" is named after the British mathematician and WWII codebreaker William Gordon Welchman, who worked at "Station X" at Bletchley Park.',
+    },
     SEB3404 = {
+        Trivia = {
+            'The octagonal sensors in shield domes are based on the SBX-1 Sea-based X-band Radar.',
+            'The largest dish in the array is based on the dish "Arthur" at Goonhilly Satellite Earth Station.',
+            'The dishes are named after Patreon supporters.',
+        },
         Videos = {
             {YouTube = 'L7nRpxNihJA', 'Time-lapse'},
         },
     },
+    SEB4102 = {
+        Trivia = {
+            [[The "LSD - Pulse" follows the naming convention of the vanilla shields 'SD' and 'HSD - Pulse', meaning, presumably, (**H**eavy) **S**hiel**d** or **S**hield **D**ome. LSD therefor meaning Light Shield Dome, and *not* lysergic acid diethylamide.]],
+            [[This was the 4th unit added to *BrewLAN*; added September 2009. It is the oldest unit to have not had a visual overhaul.]],
+        },
+    },
     SEB4303 = {
+        Trivia = "This is a recreation of an Xbox 360 exclusive unit.\n\n<blockquote>To get good reference images, I bought a copy only to find that the disk drive no longer worked on my Xbox. The reference images used were photographs of a screen taken by a friend I borrowed the game to.\n\n—Balthazar</blockquote>",
         Videos = {
             {YouTube = 'NnJHxvF0nqM', 'Time-lapse'},
         },
@@ -126,24 +212,54 @@ UnitData = {
         Videos = {
             {YouTube = '6G8ePA3qLl4', 'Time-lapse'},
         },
+        Trivia = 'BFS stands for big shield.',
+    },
+    SEB5210 = {
+        Trivia = "The appearance of this unit is loosely based on the GDI walls from Tiberian Sun. The name **G**eneral **D**estruction **I**mpasse is a nod to this.",
+    },
+    SEB9101 = {
+        Trivia = '"Tyson Campus" is named after the astrophysicist Neil deGrasse Tyson.',
+    },
+    SEB9201 = {
+        Trivia = '"Sagan Campus" is named after the cosmologist Carl Sagan.',
+    },
+    SEB9301 = {
+        Trivia = '"Hawking Campus" is named after the theoretical physicist Stephen Hawking.',
+    },
+    SEL0119 = {
+        Trivia = '"Wally" is named after WALL-E.',
+    },
+    SEL0319 = {
+        Trivia = '"Custodian" is named after an old *Total Annihilation* mod unit from *Core Prime*.',
+    },
+    SEL0320 = {
+        Trivia = [[The name "Archimedes" is a reference to the *Fallout: New Vegas* satellite weapon 'ARCHIMEDES II', other names considered include 'Athena'; a Red Alert 3 reference, and Euclid; another *F:NV* reference.]],
     },
     SEL0323 = {
+        Trivia = '"Trickshot" was named by community poll. Close second was Phalanx. Centurion wasn\'t an option, despite that being the name of the real-world land-version of the Phalanx CIWS, to avoid confusion with the [Experimental AA Gunship](SEA0401).',
         Videos = {
             {YouTube = '2b-OMEgvwFk', 'Time-lapse'},
         },
     },
     SEL0324 = {
+        Trivia = 'The top part of the Longbow is based on the Crotale EDIR.',
         Videos = {
             {YouTube = 'Ujh3mvra4ck', 'Time-lapse part 1'},
             {YouTube = 'amEokjAcKj4', 'Time-lapse part 2'},
         },
     },
     SEL0326 = {
+        Trivia = '"Bird Nest" was named by community poll. Distant second was *none of the above*, with 13% of votes.',
         Videos = {
             {YouTube = 'PWxvzfSKiFc', 'Time-lapse'},
         },
     },
     SEL3401 = {
+        Trivia = {
+            '"Teatut" is named after the mathematician and WWII codebreaker William T. Tutte, who worked at Bletchley Park.',
+            'The front radar dish is based on the AN/SPS-49.',
+            'The rear radar dish is based on the AN/FPS-6',
+        },
         Videos = {
             {YouTube = 'I63oGNkngaE', 'Time-lapse'},
         },
@@ -175,6 +291,9 @@ UnitData = {
             {YouTube = '1wDhOwlvm5w', 'Time-lapse'},
         },
     },
+    SRA0307 = {
+        Trivia = '"Zenith" was the first unit added to *BrewLAN* with an almost completely custom model.',
+    },
     SRA0313 = {
         Videos = {
             {YouTube = 'QIBbM6kEPnc', 'Time-lapse'},
@@ -189,28 +308,64 @@ UnitData = {
         Videos = {
             {YouTube = 'KI9QWzsM6QU', 'Time-lapse'},
         },
+        Trivia = '"Xiphos" was a commissioned unit. The name is from a type of ancient Greek shordsword.',
     },
     SRA0401 = {
         Videos = {
             {YouTube = 'DhKjgWez1Ao', 'Time-lapse'},
         },
+        Trivia = '"Simurgh" was a commissioned unit based on the Arsenal Bird from *Ace Combat 7*.',
     },
     SRB0401 = {
         Videos = {
             {YouTube = 'N3NnvM1SPHw', 'Time-lapse'},
         },
     },
+    SRB1311 = {
+        Trivia = 'Michael Faraday, for whom this unit was named, was an English scientist who contributed to the study of electromagnetism and electrochemistry. The basic concept of the unit is based on similar *Total Annihilation* units.',
+    },
+    SRB2103 = {
+        Trivia = [["Little Bertha" was designed from memories of the *Total Annihilation* unit 'Big Bertha', which it is also named after.]],
+    },
+    SRB9101 = {
+        Trivia = '"Einstein Facility" is named after the theoretical physicist Albert Einstein. The name Oppenheimer was also considered.',
+    },
+    SRB9201 = {
+        Trivia = '"Schrodinger Facility" is named after the cat that physicist Erwin Schrödinger may or may not have owned.',
+    },
+    SRB9301 = {
+        Trivia = '"Heisenberg Facility" is named after the theoretical physicist Werner Heisenberg.',
+    },
+    SRL0119 = {
+        Trivia = '"Outy" takes its name from the username of one of the first play testers of *BrewLAN*.',
+    },
+    SRL0209 = {
+        Trivia = '"Wilson" is named after one of the first play testers of *BrewLAN* who mained Cybran and liked to build walls.',
+    },
     SRL0311 = {
+        Trivia = {
+            [[The name "Triseptitron" is based on the Latin cardinal numbers for 3 and 7; tri and septi, referring to the 3 racks of 7 projectiles.]],
+            [[Prior to receiving a new model, this was called "Hexatron". 'Hexa', the Greek cardinal for 6, referring to honeycomb muzzle pattern.]],
+            [[A more linguistically consistent name would be 'Triheptatron'; the Greek cardinal prefixes for 3 and 7, to match the Greek -tron suffix.]],
+        },
         Videos = {
             {YouTube = 'tzP_zOwRQGA', 'Time-lapse'},
         },
     },
     SRL0312 = {
+        Trivia = {
+            '"Toxotai" was a commissioned unit.',
+            'The name "Toxotai" is the plural form of a form of an Ancient Greek/Byzantine archer armed with a shortsword. The singular would be "Toxotes".',
+        },
         Videos = {
             {YouTube = 'clowfXPczHc', 'Time-lapse'},
         },
     },
+    SRL0319 = {
+        Trivia = '"House" takes its name from the eponymous character from the TV show of the same name. It was chosen as a red herring for the origin of the name of "[Wilson](SRL0209)".',
+    },
     SRL0320 = {
+        Trivia = '"Slink" is loosely based on the Nod stealth tank.',
         Videos = {
             {YouTube = 'LVfU06B7kv4', 'Time-lapse'},
         },
@@ -233,8 +388,10 @@ UnitData = {
     SRB2391 = {
         AdjacencySuffix = "Adjacency effects that modify the maintenance cost will proportionally change the Tesla weapon damage as long as the net effect is an increase, not a decrease. Meaning that basic power adjacency will lower the maintenance cost, but not alter the damage.",
         WeaponsPrefix = "The Tesla weapon has a lifetime sustained DPS of 40 regardless of how often it fires. For every second it isn't firing it stores 40 damage, and it releases all that damage with every shot. There is no upper limit to the damage per shot, and the effective minimum damage per shot is 28 based on the firing cycle of once every 0.7 seconds.\n\nThe death weapon deals 10 damage for every second since the Tesla weapon last fired.",
+        Trivia = 'The health and unit ID of "Tesla Coil" are both references to the year the Tesla coil was invented; 1981.',
     },
     SRB3301 = {
+        Trivia = '"Qain", most directly, takes its name from the tech YouTuber [Qain](https://youtube.com/deusqain); a self professed technomancer and cyborg, but also partially from the films *Citizen Kane*, and *Raising Cain*, and the Biblical figure Cain.',
         Videos = {
             {YouTube = 'pJ6C0kDOGhM', 'Time-lapse'},
         },
@@ -242,7 +399,10 @@ UnitData = {
 
 
     SSA0001 = {
-        LeadSuffix = "When <a href=\"SSL0403\">Iyadesu</a> reclaims a wreck, it creates one of these drones to store the blueprint of said wreck.",
+        LeadSuffix = "When [Iyadesu](SSL0403) reclaims a wreck, it creates one of these drones to store the blueprint of said wreck.",
+    },
+    SSA0306 = {
+        Trivia = '"Vishuum" was one of the handful of units that predates the first public release of *BrewLAN*. It is one of the few units from that era to have remained largely unchanged.',
     },
     SSA0313 = {
         Videos = {
@@ -265,29 +425,35 @@ UnitData = {
         },
     },
     SSB4401 = {
-        LeadSuffix = "When the shield activates it creates a lattice of shield drone units that orbit the main structure. The lattice consists of an inner ring of 6 \"<a href=\"SSB4401_Large\">Iaathanah</a>\", and an outer ring of 12 \"<a href=\"SSB4401_Small\">Iaathanahee</a>\". The approximate total coverage area is represented by the misc overlay ring.",
+        LeadSuffix = 'When the shield activates it creates a lattice of shield drone units that orbit the main structure. The lattice consists of an inner ring of 6 "[Iaathanah](SSB4401_Large)", and an outer ring of 12 "[Iaathanahee](SSB4401_Small)". The approximate total coverage area is represented by the misc overlay ring.',
     },
     SSB4401_Large = {
-        LeadSuffix = "It is created as part of the shield activation of \"<a href=\"SSB4401\">Iaathan</a>\" alongside \"<a href=\"SSB4401_Small\">Iaathanahee</a>\".",
+        LeadSuffix = 'It is created as part of the shield activation of "[Iaathan](SSB4401)" alongside "[Iaathanahee](SSB4401_Small)".',
     },
     SSB4401_Small = {
-        LeadSuffix = "It is created as part of the shield activation of \"<a href=\"SSB4401\">Iaathan</a>\" alongside \"<a href=\"SSB4401_Large\">Iaathanah</a>\".",
+        LeadSuffix = 'It is created as part of the shield activation of "[Iaathan](SSB4401)" alongside "[Iaathanah](SSB4401_Large)".',
+    },
+    SSB5401 = {
+        Trivia = [["Chappa'ai" is based on the gate design used in *Stargate Universe*. The name is the Goa'uld name for Stargate. If loaded without localisation, it will instead have the name 'Chappa'ko'; a reference to season 9‒10 *Stargate SG1*.]],
     },
     SSL0403 = {
         Videos = {
             {YouTube = 'IInITjdtaPM', 'Time-lapse'},
         },
-        LeadSuffix = "Upon completing reclamation, it gains a <a href=\"SSA0001\">reconstruction drone</a> that stores the associated blueprint. Both it and the drone can build said blueprint and the blueprint is lost if the drone dies. Redundant drones can be used to protect blueprints.",
+        LeadSuffix = "Upon completing reclamation, it gains a [reconstruction drone](SSA0001) that stores the associated blueprint. Both it and the drone can build said blueprint and the blueprint is lost if the drone dies. Redundant drones can be used to protect blueprints.",
+        Trivia = 'Several props have effects when reclaimed by the "Iyadesu".',
     },
     SSL0404 = {
         Videos = {
             {YouTube = 'tYGZTTVA5lQ', 'Time-lapse'},
         },
+        Trivia = '"Sisha-Ah" was a commissioned unit.',
     },
     SSL0405 = {
         Videos = {
             {YouTube = '9E--KMyfdBI', 'Trailer video'},
             {YouTube = 'aOUeAy24ong', 'Time-lapse'},
         },
+        Trivia = '"Suedath-Zmara" was commissioned to commemorate the death of a member of the LOUD team, Henry Thaddeus Katzmarek. Numerous aspects of the unit reference this fact: The name takes part of his name backwards, the health and regen reference his birth and death years; 1959‒2020, and the energy cost references the date of death; 1st June 2020.',
     },
 }